changing serial format to use 32 bit numbers

+++ cpp/lib/extend/interface.hpp
@@ -17,8 +17,8 @@ struct PackedFmt {
     void pack(const char *query, const char *subject, u64 query_size, u64 subject_size);
 
     u64 query_size;
-    u64 gap_index;
-    u64 gap_count;
+    u32 gap_index;
+    u32 gap_count;
     byte query[SW_MAX_LENGTH];
     byte subject[SW_MAX_LENGTH];
 };
